NSSummaryTask abstract class
